+++ Tools/MinimalDeploy.tests.ps1
@@ -46,6 +46,16 @@ Describe "Deploys Minimal API" {
         Test-MainPod $deploy.PodStatuses[0]
     } -Tag 'Happy'
 
+    It "runs with prehook only ok" {
+        $deploy = Deploy-Minimal -PassThru  -SkipInit -SkipDeploy
+
+        $deploy.Running | Should -Be $false
+        $deploy.ReleaseName | Should -Be 'test'
+        $deploy.PodStatuses | Should -Be $null
+        $deploy.PreHookStatus.Status | Should -Be 'Running'
+        $deploy.RollbackStatus | Should -Be 'DeployedOk'
+    } -Tag 'Happy'
+
     It "runs a dry run" {
         $deploy = Deploy-Minimal -PassThru -DryRun 2>&1 | Out-Null
         $deploy | Should -Be $null
@@ -53,28 +63,28 @@ Describe "Deploys Minimal API" {
 
     It "has main container crash" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -Fail
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Crash'
     } -Tag 'Crash','Negative'
 
     It "has main container has bad image tag" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -ImageTag zzz
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'ConfigError' -reason "ErrImageNeverPull"
     } -Tag 'Config','Negative'
 
     It "has the main container with a bad secret name" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -BadSecret
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'ConfigError' -reason "CreateContainerConfigError"
     } -Tag 'Config','Negative'
 
     It "has the main container time out" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -RunCount 100
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
     } -Tag 'Timeout','Negative'
@@ -88,14 +98,14 @@ Describe "Deploys Minimal API" {
 
     It "has a temporary startup timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 10 -RunCount 10 -StartupProbe
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
     } -Tag 'Negative', 'Timeout', 'Probe'
 
     It "has a bad probe" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -SkipPreHook -TimeoutSec 120 -Readiness '/fail'
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
         $deploy.PodStatuses[0].LastBadEvents.Count | Should -Be 1
@@ -104,14 +114,14 @@ Describe "Deploys Minimal API" {
 
     It "has a prehook job top timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -HookRunCount 50 -TimeoutSec 10
-        Test-Deploy $deploy -Running $false -podCount 0
+        Test-Deploy $deploy -Running $false -podCount 0 -RollbackStatus 'RolledBack'
 
         # no pod statuses
     } -Tag 'Negative', 'Timeout'
 
     It "has an init timeout" {
         $deploy = Deploy-Minimal -PassThru -SkipPreHook -TimeoutSec 10 -InitRunCount 50
-        Test-Deploy $deploy -Running $false
+        Test-Deploy $deploy -Running $false -RollbackStatus 'RolledBack'
 
         Test-MainPod $deploy.PodStatuses[0] -status 'Unknown' -reason "Possible timeout"
     } -Tag 'Negative', 'Timeout'
@@ -119,6 +129,7 @@ Describe "Deploys Minimal API" {
     It "has the prehook job hook times" {
         $deploy = Deploy-Minimal -PassThru -SkipInit -HookRunCount 100 -PreHookTimeoutSecs 5
         $deploy.Running | Should -Be $false
+        $deploy.RollbackStatus | Should -Be 'RolledBack'
         $deploy.ReleaseName | Should -Be 'test'
 
         Test-PreHook $deploy.PreHookStatus # will be running
@@ -126,8 +137,8 @@ Describe "Deploys Minimal API" {
 
     It "has prehook job crash" {
         $deploy = Deploy-Minimal -PassThru -HookFail -TimeoutSecs 20 -PreHookTimeoutSecs 20
-        Test-Deploy $deploy -Running $false -PodCount 0
+        Test-Deploy $deploy -Running $false -PodCount 0 -RollbackStatus 'RolledBack'
         $deploy.PreHookStatus.Status | Should -Be 'Crash'
-    } -Tag 'Crash','Negative','New'
+    } -Tag 'Crash','Negative'
 }